+++ python/acl_anthology/people/index.py
@@ -213,6 +213,52 @@ class PersonIndex(SlottedDict[Person]):
         del coauthors[person.id]
         return coauthors
 
+    def generate_person_id(
+        self,
+        person_or_name: Person | Name,
+        suffix: Optional[str] = None,
+        orcid: Optional[str] = None,
+    ) -> str:
+        """Generate a verified person ID that does not exist yet in the index.
+
+        This does _not_ modify the index or a supplied Person instance in any way.
+
+        Parameters:
+            person_or_name: A Person or Name instance to use as the basis for generating the ID.  If given a Person, their canonical name will be used, and potentially their ORCID iD (if set).
+            suffix: A suffix to append to the ID in case the generated ID is already taken.
+            orcid: An ORCID iD whose last four characters can be appended to the ID in case the generated ID is already taken.  If None and the first argument is a Person, the Person's ORCID iD will be used instead.
+
+        Returns:
+            A valid verified person ID that doesn't exist yet.
+
+        Raises:
+            AnthologyException: If all generated IDs already exist in the index.
+        """
+        if not self.is_data_loaded:
+            self.load()
+
+        if isinstance(person_or_name, Person):
+            slug = person_or_name.canonical_name.slugify()
+        elif isinstance(person_or_name, Name):
+            slug = person_or_name.slugify()
+        else:  # pragma: no cover
+            raise TypeError(
+                f"First argument must be Person or Name (got '{person_or_name!r}' that is a {type(person_or_name)!r})"
+            )
+
+        if slug not in self.data:
+            return slug
+
+        if suffix is not None and (pid := f"{slug}-{suffix.lower()}") not in self.data:
+            return pid
+
+        if orcid is None and isinstance(person_or_name, Person):
+            orcid = person_or_name.orcid
+        if orcid is not None and (pid := f"{slug}-{orcid[-4:].lower()}") not in self.data:
+            return pid
+
+        raise AnthologyException(f"Could not generate an ID for {person_or_name!r}")
+
     def load(self) -> None:
         """Loads or builds the index."""
         # This function exists so we can later add the option to read the index
@@ -483,11 +529,8 @@ class PersonIndex(SlottedDict[Person]):
             # Make sure the name used here is listed for this person
             person.add_name(name_spec.name)
         else:
-            # Need to create a new person; generate name slug for the ID
-            pid = name_spec.name.slugify()
-            if pid in self.data:
-                # ID is already in use; add last four digits of ORCID to disambiguate
-                pid = f"{pid}-{name_spec.orcid[-4:].lower()}"
+            # Need to create a new person; generate ID
+            pid = self.generate_person_id(name_spec.name, orcid=name_spec.orcid)
 
             self.add_person(
                 Person(