+++ python/tests/people/person_test.py
@@ -209,7 +209,20 @@ def test_person_is_explicit(anthology):
 def test_person_make_explicit(anthology):
     person = anthology.get_person(UNVERIFIED_PID_FORMAT.format(pid="nicoletta-calzolari"))
     assert not person.is_explicit
-    person.make_explicit("nicoletta-calzolari")
+    person.make_explicit("nicoletta-calzolari-abc")
+    assert person.is_explicit
+    assert person.id == "nicoletta-calzolari-abc"
+    # IDs have been added to these collections:
+    assert anthology.collections["J89"].is_modified
+    assert anthology.collections["L06"].is_modified
+    # But not this one:
+    assert not anthology.collections["2022.acl"].is_modified
+
+
+def test_person_make_explicit_with_default_id(anthology):
+    person = anthology.get_person(UNVERIFIED_PID_FORMAT.format(pid="nicoletta-calzolari"))
+    assert not person.is_explicit
+    person.make_explicit()  # default ID
     assert person.is_explicit
     assert person.id == "nicoletta-calzolari"
     # IDs have been added to these collections:
@@ -219,11 +232,48 @@ def test_person_make_explicit(anthology):
     assert not anthology.collections["2022.acl"].is_modified
 
 
+def test_person_make_explicit_skip_setting_ids(anthology):
+    person = anthology.get_person(UNVERIFIED_PID_FORMAT.format(pid="nicoletta-calzolari"))
+    assert not person.is_explicit
+    person.make_explicit(new_id="nicoletta-calzolari", skip_setting_ids=True)
+    assert person.is_explicit
+    assert person.id == "nicoletta-calzolari"
+    # IDs have NOT been added to these collections:
+    assert not anthology.collections["J89"].is_modified
+    assert not anthology.collections["L06"].is_modified
+    assert not anthology.collections["2022.acl"].is_modified
+
+
 def test_person_make_explicit_should_raise_when_explicit(anthology):
     person = anthology.get_person("marcel-bollmann")
     assert person.is_explicit
     with pytest.raises(AnthologyException):
-        person.make_explicit("marcel-bollmann")
+        person.make_explicit("marcel-bollmann-new")
+
+
+def test_person_set_id_on_items(anthology):
+    person = anthology.get_person("steven-krauwer")
+    # Verify precondition: Not all papers associated with this person have an ID set
+    namespecs = list(person.namespecs())
+    assert any(namespec.id is None for namespec in namespecs)
+    # Set IDs, excluding one item by ID
+    person.set_id_on_items(exclude=["2022.naloma-1.9"])
+    paper = anthology.get_paper("2022.naloma-1.9")
+    assert any(namespec.id is None for namespec in namespecs)
+    assert paper.authors[0].id is None
+    person.set_id_on_items(exclude=[paper])
+    assert any(namespec.id is None for namespec in namespecs)
+    assert paper.authors[0].id is None
+    # Set ID, not excluding the paper
+    person.set_id_on_items()
+    assert all(namespec.id == "steven-krauwer" for namespec in namespecs)
+    assert paper.authors[0].id == "steven-krauwer"
+
+
+def test_person_set_id_on_items_should_raise(anthology):
+    person = anthology.get_person(UNVERIFIED_PID_FORMAT.format(pid="nicoletta-calzolari"))
+    with pytest.raises(AnthologyException):
+        person.set_id_on_items()
 
 
 def test_person_merge_into_unverified_verified(anthology):