+++ python/CHANGELOG.md
@@ -6,8 +6,10 @@
 
 - Files (such as paper PDFs) can now be downloaded from their remote URLs via `.download()`.
 - Added `.get_namespec_for(Person)` on papers and volumes, to more easily find the NameSpecification referring to a given Person.
-- Added `Person.namespecs()` to iterate over all NameSpecifications referring to this Person.
 - Added `NameSpecification.case_normalize()` to heuristically fix casing and match spelling details to known canonical names. This is mainly intended for ingestion and called automatically when using `create_paper()`/`create_volume()`.
+- Added `Person.namespecs()` to iterate over all NameSpecifications referring to this Person.
+- Added `Person.set_id_on_items()` to explicitly set a verified person's ID on all NameSpecifications that currently resolve to them.
+- Added `PersonIndex.generate_person_id()` to facilitate generating verified person IDs that don't exist yet in the index.
 
 ### Changed
 
@@ -18,6 +20,7 @@
   - NameSpecifications now track if they have been modified, and will trigger their parent collection being saved on `Anthology.save_all()`.
   - NameSpecifications can now be resolved via `NameSpecification.resolve()`. Therefore, `Anthology.resolve()` has been deprecated.
 - Improvements to Person:
+  - `Person.make_explicit()` can now auto-generate an ID.
   - `Person.merge_with_explicit()` has been renamed `Person.merge_into()` and now supports merging two explicit persons.
 
 ## [1.0.0] — 2026-01-24